+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.5.0
+### Features
+* add global error handler
+
+## 0.4.2
+### Bugfix
+* fix testing framework when parsing expression with inline functions
+
 ## 0.4.1
 ### Bugfix
 * fix false warning about not triggered handlers on input events